+++ docs/src/pages/[platform]/connected-components/storage/fileuploader/examples/VariationExample.tsx
@@ -1,0 +1,19 @@
+import { FileUploader } from '@aws-amplify/ui-react';
+export const VariationExample = () => {
+  return (
+    <>
+      <FileUploader
+        variation="drop"
+        acceptedFileTypes={['image/*']}
+        accessLevel="public"
+        provider="fast" // IGNORE
+      />
+      <FileUploader
+        variation="button"
+        acceptedFileTypes={['image/*']}
+        accessLevel="public"
+        provider="fast" // IGNORE
+      />
+    </>
+  );
+};