+++ examples/react-native/metro.config.js
@@ -103,10 +103,6 @@ config.resolver.extraNodeModules = {
   '@xstate/react': path.resolve(__dirname, 'node_modules/@xstate/react'),
 
   // depedendencies with native modules
-  '@react-native-clipboard/clipboard': path.resolve(
-    __dirname,
-    'node_modules/@react-native-clipboard/clipboard'
-  ),
   '@react-native-picker/picker': path.resolve(
     __dirname,
     'node_modules/@react-native-picker/picker'