+++ examples/react-native/src/features/Authenticator/Styles/Example.tsx
@@ -1,12 +1,26 @@
 import React, { ReactNode } from 'react';
-import { StyleProp, StyleSheet, Text, View, ViewStyle } from 'react-native';
+import {
+  Button,
+  StyleProp,
+  StyleSheet,
+  Text,
+  useColorScheme,
+  View,
+  ViewStyle,
+} from 'react-native';
 
-import { Authenticator, useAuthenticator } from '@aws-amplify/ui-react-native';
+import {
+  Authenticator,
+  useAuthenticator,
+  defaultDarkModeOverride,
+  ThemeProvider,
+  Theme,
+  useTheme,
+} from '@aws-amplify/ui-react-native';
 import { Amplify } from 'aws-amplify';
 
-import { Button } from '../../../ui';
-
-Amplify.configure({});
+import awsconfig from './aws-exports';
+Amplify.configure(awsconfig);
 
 const MyHeader = ({
   children,
@@ -14,32 +28,48 @@ const MyHeader = ({
 }: {
   children?: ReactNode;
   style?: StyleProp<ViewStyle>;
-}) => (
-  <View style={style}>
-    <Text style={{ fontSize: 36 }}>{children}</Text>
-  </View>
-);
+}) => {
+  const {
+    tokens: { colors, fontSizes },
+  } = useTheme();
+  return (
+    <View style={style}>
+      <Text
+        style={{ fontSize: fontSizes.xxl, color: colors.brand.primary[80] }}
+      >
+        {children}
+      </Text>
+    </View>
+  );
+};
 
 function SignOutButton() {
   const { signOut } = useAuthenticator();
-  return <Button onPress={signOut}>Sign Out</Button>;
+  return <Button onPress={signOut} title="Sign Out" />;
 }
 
+const theme: Theme = {
+  overrides: [defaultDarkModeOverride],
+};
+
 function App() {
+  const colorMode = useColorScheme();
   return (
-    <Authenticator.Provider>
-      <Authenticator
-        components={{
-          SignIn: (props) => (
-            <Authenticator.SignIn {...props} Header={MyHeader} />
-          ),
-        }}
-      >
-        <View style={style.container}>
-          <SignOutButton />
-        </View>
-      </Authenticator>
-    </Authenticator.Provider>
+    <ThemeProvider theme={theme} colorMode={colorMode}>
+      <Authenticator.Provider>
+        <Authenticator
+          components={{
+            SignIn: (props) => (
+              <Authenticator.SignIn {...props} Header={MyHeader} />
+            ),
+          }}
+        >
+          <View style={style.container}>
+            <SignOutButton />
+          </View>
+        </Authenticator>
+      </Authenticator.Provider>
+    </ThemeProvider>
   );
 }