+++ packages/react-native/src/Authenticator/Defaults/SetupTOTP/styles.ts
@@ -1,33 +1,10 @@
-import {
-  ImageStyle,
-  StyleProp,
-  StyleSheet,
-  TextStyle,
-  ViewStyle,
-} from 'react-native';
+import { StyleSheet, TextStyle } from 'react-native';
 
 export interface SetupTOTPStyle {
-  buttonPrimary: ViewStyle;
-  buttonSecondary: ViewStyle;
-  secretKeyContainer: ViewStyle;
   secretKeyText: TextStyle;
-  copyIcon: StyleProp<ImageStyle>;
 }
 
 export const styles: SetupTOTPStyle = StyleSheet.create({
-  buttonPrimary: {
-    marginVertical: 8,
-  },
-  buttonSecondary: {
-    marginVertical: 8,
-  },
-  secretKeyContainer: {
-    alignItems: 'center',
-    flexDirection: 'row',
-    paddingVertical: 8,
-  },
-  secretKeyText: { flex: 1, flexWrap: 'wrap' },
-  copyIcon: {
-    marginLeft: 12,
-  },
+  // TODO: replace with tokens
+  secretKeyText: { flexWrap: 'wrap', paddingVertical: 4, marginHorizontal: 8 },
 });