+++ packages/react-native/src/Authenticator/common/DefaultContent/styles.ts
@@ -4,30 +4,54 @@ import { StrictTheme } from '../../../theme';
 import { DefaultContentStyle } from './types';
 
 export const getDefaultStyle = ({
-  tokens: { space },
+  tokens: { colors, fontSizes, space },
 }: StrictTheme): Required<DefaultContentStyle> =>
   StyleSheet.create({
+    body: {
+      fontSize: fontSizes.medium,
+      paddingHorizontal: space.small,
+    },
     buttonPrimary: {
-      marginVertical: space.medium,
+      marginHorizontal: space.medium,
+      marginVertical: space.small,
     },
     buttonPrimaryLabel: {}, // themed value only
     buttonSecondary: {
+      marginHorizontal: space.medium,
+    },
+    buttonSecondaryLabel: {}, // themed value only
+    errorMessage: {
+      marginVertical: space.small,
+    },
+    errorMessageLabel: {}, // themed value only
+    fieldContainer: {
+      paddingHorizontal: space.small,
+    }, // themed value only
+    fieldErrorsContainer: {
+      paddingHorizontal: space.medium,
+      paddingVertical: space.xs,
+    },
+    fieldError: {
+      color: colors.font.error,
+      paddingVertical: 1,
+    }, // themed value only
+    fieldLabel: {}, // themed value only
+    fieldStyle: {}, // themed value only
+    footer: {}, // themed value only
+    formFields: {
+      paddingBottom: space.xs,
+    },
+    header: {
+      marginVertical: space.small,
+      paddingHorizontal: space.small,
+    },
+    link: {
       marginVertical: space.xs,
       minWidth: '50%',
     },
-    buttonSecondaryContainer: {
+    linksContainer: {
       flexDirection: 'row',
       flexWrap: 'wrap',
       justifyContent: 'center',
     },
-    buttonSecondaryLabel: {}, // themed value only
-    errorMessage: { marginVertical: space.small },
-    errorMessageLabel: {}, // themed value only
-    fieldContainerStyle: {}, // themed value only
-    fieldErrorStyle: {}, // themed value only
-    fieldLabelStyle: {}, // themed value only
-    fieldStyle: {}, // themed value only
-    footer: { height: 24, width: '100%' },
-    formFields: { marginVertical: space.small },
-    header: { marginVertical: space.small },
   });