+++ packages/react-native/src/Authenticator/common/DefaultContent/types.ts
@@ -3,9 +3,10 @@ import { StyleProp, TextStyle, ViewStyle } from 'react-native';
 
 import {
   AuthenticatorComponentDefaultProps,
-  AuthenticatorRouteComponentName,
   AuthenticatorHeaderComponent,
   AuthenticatorFooterComponent,
+  AuthenticatorMachineContext,
+  AuthenticatorRouteComponentName,
 } from '@aws-amplify/ui-react-core';
 
 import { ButtonProps } from '../../../primitives';
@@ -17,20 +18,23 @@ import {
 import { TextFieldOptionsType, RadioFieldOptions } from '../../hooks';
 
 export interface DefaultContentStyle {
+  body?: TextStyle;
   buttonPrimary?: ViewStyle;
   buttonPrimaryLabel?: TextStyle;
   buttonSecondary?: ViewStyle;
-  buttonSecondaryContainer?: ViewStyle;
   buttonSecondaryLabel?: TextStyle;
   errorMessage?: ViewStyle;
   errorMessageLabel?: TextStyle;
-  fieldContainerStyle?: ViewStyle;
-  fieldErrorStyle?: TextStyle;
-  fieldLabelStyle?: TextStyle;
+  fieldErrorsContainer?: ViewStyle;
+  fieldContainer?: ViewStyle;
+  fieldError?: TextStyle;
+  fieldLabel?: TextStyle;
   fieldStyle?: TextStyle;
   footer?: TextStyle;
   formFields?: ViewStyle;
   header?: TextStyle;
+  link?: ViewStyle;
+  linksContainer?: ViewStyle;
 }
 
 export type DefaultContentStyleProps = {
@@ -46,7 +50,8 @@ type AuthenticatorButtonProps = Omit<ButtonProps, 'children'> & {
 
 type DefaultButtons = {
   primary: AuthenticatorButtonProps;
-  secondary: AuthenticatorButtonProps | AuthenticatorButtonProps[];
+  secondary?: AuthenticatorButtonProps;
+  links?: AuthenticatorButtonProps[];
 };
 
 export type DefaultContentProps<
@@ -61,8 +66,9 @@ export type DefaultContentProps<
   buttons: DefaultButtons;
   body?: React.ReactNode;
   fields: FieldsType[];
-  headerText?: string;
+  headerText: string;
   Footer: AuthenticatorFooterComponent<{ style?: StyleProp<TextStyle> }>;
   FormFields: DefaultFormFieldsComponent<FieldsType>;
   Header: AuthenticatorHeaderComponent<{ style?: StyleProp<TextStyle> }>;
+  validationErrors?: AuthenticatorMachineContext['validationErrors'];
 };