+++ packages/react-native/src/primitives/PasswordField/styles.ts
@@ -3,13 +3,18 @@ import { StyleSheet } from 'react-native';
 import { StrictTheme } from '../../theme';
 import { PasswordFieldStyles } from './types';
 
-export const getThemedStyles = (theme: StrictTheme): PasswordFieldStyles => {
-  const { components } = theme;
-
+export const getThemedStyles = ({
+  components,
+  tokens: { colors },
+}: StrictTheme): PasswordFieldStyles => {
   return StyleSheet.create({
     container: {
       ...components?.passwordField?.container,
     },
-    icon: { ...components?.passwordField?.icon },
+    icon: {
+      // match to `TextField` placeholder color
+      tintColor: colors.font.tertiary,
+      ...components?.passwordField?.icon,
+    },
   });
 };