+++ packages/react/src/primitives/Autocomplete/__tests__/Autocomplete.test.tsx
@@ -65,6 +65,13 @@ describe('Autocomplete: ', () => {
     userEvent.keyboard('{Esc}');
     expect(textInput).toHaveValue('');
 
+    // Focus will not open menu
+    userEvent.click(document.body);
+    expect(textInput).not.toHaveFocus();
+    textInput.focus();
+    expect(textInput).toHaveFocus();
+    expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+
     // Click will open menu
     expect(textInput).toHaveFocus();
     userEvent.click(textInput);
@@ -188,7 +195,7 @@ describe('Autocomplete: ', () => {
     render(<Autocomplete label={label} options={options} isLoading />);
 
     const textInput = await screen.findByRole('combobox');
-    textInput.focus();
+    userEvent.click(textInput);
     const listbox = screen.queryByRole('listbox');
     expect(listbox).not.toBeInTheDocument();
     const loading = screen.getByText(ComponentText.Autocomplete.loadingText);
@@ -199,7 +206,7 @@ describe('Autocomplete: ', () => {
     render(<Autocomplete label={label} options={options} />);
 
     const textInput = await screen.findByRole('combobox');
-    textInput.focus();
+    userEvent.click(textInput);
 
     const appleOption = screen.getByText('apple')
       .parentElement as HTMLLIElement;
@@ -255,6 +262,21 @@ describe('Autocomplete: ', () => {
     );
   });
 
+  it('should be case insensitive filtering', async () => {
+    render(<Autocomplete label={label} options={options} />);
+
+    const textInput = await screen.findByRole('combobox');
+    userEvent.type(textInput, 'ap');
+    let optionElements = await screen.findAllByRole('option');
+    expect(optionElements).toHaveLength(1);
+    expect(optionElements[0]).toHaveTextContent('apple');
+    userEvent.clear(textInput);
+    userEvent.type(textInput, 'AP');
+    optionElements = await screen.findAllByRole('option');
+    expect(optionElements).toHaveLength(1);
+    expect(optionElements[0]).toHaveTextContent('apple');
+  });
+
   it('should be able to apply custom filtering', async () => {
     const optionFilter = jest.fn();
     render(
@@ -281,7 +303,7 @@ describe('Autocomplete: ', () => {
     );
 
     const textInput = await screen.findByRole('combobox');
-    textInput.focus();
+    userEvent.click(textInput);
     expect(renderOption).toHaveBeenCalled();
   });
 
@@ -297,7 +319,7 @@ describe('Autocomplete: ', () => {
     );
 
     const textInput = await screen.findByRole('combobox');
-    textInput.focus();
+    userEvent.click(textInput);
 
     const loading = screen.getByText(LoadingIndicator);
     expect(loading).toBeInTheDocument();
@@ -309,7 +331,7 @@ describe('Autocomplete: ', () => {
     render(<Autocomplete label={label} options={[]} menuSlots={{ Empty }} />);
 
     const textInput = await screen.findByRole('combobox');
-    textInput.focus();
+    userEvent.click(textInput);
 
     const empty = screen.getByText(Empty);
     expect(empty).toBeInTheDocument();
@@ -328,7 +350,7 @@ describe('Autocomplete: ', () => {
     );
 
     const textInput = await screen.findByRole('combobox');
-    textInput.focus();
+    userEvent.click(textInput);
 
     const header = screen.getByText(Header);
     expect(header).toBeInTheDocument();
@@ -347,7 +369,7 @@ describe('Autocomplete: ', () => {
     expect(textInput).toHaveAttribute('aria-haspopup', 'listbox');
     expect(textInput).toHaveAttribute('aria-expanded', 'false');
 
-    textInput.focus();
+    userEvent.click(textInput);
     const menu = screen.getByRole('listbox').parentElement;
     expect(textInput).toHaveAttribute('aria-expanded', 'true');
     expect(textInput).toHaveAttribute('aria-controls', menu?.id);